05-automacao-relatorios-python/requirements.txt
pandas==2.2.3
reportlab==4.2.5
python-dotenv==1.0.1
openpyxl==3.1.5
